requirements.txt
altgraph==0.17.5
asgiref==3.11.1
attrs==25.4.0
bottle==0.13.4
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
flatbuffers==25.12.19
idna==3.11
ImageIO==2.37.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.46.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
onnxruntime==1.24.2
packaging==26.0
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
protobuf==6.33.5
proxy_tools==0.1.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
PyMatting==1.1.15
pywebview==6.1
referencing==0.37.0
rembg==2.0.72
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==82.0.0
sqlparse==0.5.5
sympy==1.14.0
tifffile==2026.2.20
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
